Tidy some Eclipse warnings
git-svn-id: https://svn-us.apache.org/repos/asf/commons/proper/jexl/branches/2.0@800809 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/jexl/util/ListGetExecutor.java
@@ -50,9 +50,9 @@ public final class ListGetExecutor extends AbstractExecutor.Get {
     @Override
     public Object execute(final Object list) {
         if (method == ARRAY_GET) {
-            return java.lang.reflect.Array.get(list, property);
+            return java.lang.reflect.Array.get(list, property.intValue());
         } else {
-            return ((List<?>) list).get(property);
+            return ((List<?>) list).get(property.intValue());
         }
     }
 
@@ -63,9 +63,9 @@ public final class ListGetExecutor extends AbstractExecutor.Get {
             && objectClass.equals(list.getClass())
             && index instanceof Integer) {
             if (method == ARRAY_GET) {
-                return java.lang.reflect.Array.get(list, property);
+                return java.lang.reflect.Array.get(list, property.intValue());
             } else {
-                return ((List<?>) list).get(property);
+                return ((List<?>) list).get(property.intValue());
             }
         }
         return TRY_FAILED;
